+++ test/stats/frequencies.js
@@ -4,11 +4,15 @@
  *
  */
 import assert from 'assert';
-import {absolute, relative} from '../../src/stats/frequencies';
+import {
+  absolute,
+  relative,
+  updateFrequencies
+  } from '../../src/stats/frequencies';
 
 describe('frequencies', function() {
 
-  describe('absolute', function() {
+  describe('#.absolute', function() {
     it('should compute correct frequencies of the given sequence.', function() {
       assert.deepEqual(
         absolute([1, 2, 3, 3, 4, 4, 4, 5]),
@@ -35,7 +39,7 @@ describe('frequencies', function() {
     });
   });
 
-  describe('relative', function() {
+  describe('#.relative', function() {
     it('should compute correct frequencies of the given sequence.', function() {
       assert.deepEqual(
         relative([1, 2, 3, 3, 4, 4, 4, 5]),
@@ -58,6 +62,41 @@ describe('frequencies', function() {
           s: 0.25
         }
       );
+    });
+
+    it('should be possible to pass absolute frequencies.', function() {
+      assert.deepEqual(
+        relative(absolute([1, 2, 3, 3, 4, 4, 4, 5])),
+        {
+          1: 1 / 8,
+          2: 1 / 8,
+          3: 2 / 8,
+          4: 3 / 8,
+          5: 1 / 8
+        }
+      );
+    });
+  });
+
+  describe('#.updateFrequencies', function() {
+
+    it('should correctly update frequencies with the given sequence.', function() {
+      const previousFrequencies = {
+        1: 1,
+        2: 1,
+        3: 2,
+        4: 3,
+        5: 1
+      };
+
+      assert.deepEqual(updateFrequencies(previousFrequencies, [7, 1, 1, 1, 2]), {
+        1: 4,
+        2: 2,
+        3: 2,
+        4: 3,
+        5: 1,
+        7: 1
+      });
     });
   });
 });